requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.13.1
google==3.0.0
google-auth==2.27.0
idna==3.6
inflection==0.5.1
packaging==23.2
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
rsa==4.9
soupsieve==2.5
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.0
virtualenv==20.25.0